+++ less/utils.py
@@ -1,4 +1,4 @@
-from less.settings import LESS_EXECUTABLE, LESS_OUTPUT_DIR
+from less.settings import LESS_EXECUTABLE, LESS_ROOT, LESS_OUTPUT_DIR
 from django.conf import settings
 import logging
 import posixpath
@@ -10,7 +10,6 @@ import subprocess
 logger = logging.getLogger("less")
 
 
-STATIC_ROOT = getattr(settings, "STATIC_ROOT", getattr(settings, "STATIC_ROOT"))
 STATIC_URL = getattr(settings, "STATIC_URL", getattr(settings, "MEDIA_URL"))
 
 
@@ -36,7 +35,7 @@ class URLConverter(object):
 
 def compile_less(input, output, less_path):
 
-    less_root = os.path.join(STATIC_ROOT, LESS_OUTPUT_DIR)
+    less_root = os.path.join(LESS_ROOT, LESS_OUTPUT_DIR)
     if not os.path.exists(less_root):
         os.makedirs(less_root)
 
@@ -51,11 +50,11 @@ def compile_less(input, output, less_path):
     output_directory = os.path.dirname(output)
     if not os.path.exists(output_directory):
         os.makedirs(output_directory)
-    compiled_file = open(output, "w+")
     compiled_css = URLConverter(
         out.decode(settings.FILE_CHARSET),
         os.path.join(STATIC_URL, less_path)
     ).convert()
+    compiled_file = open(output, "w+")
     compiled_file.write(compiled_css.encode(settings.FILE_CHARSET))
     compiled_file.close()